[BH-2085] Set lower minimum brightness for Pro version
* Added ENABLE_VERSION_PRO build flag that
selects different config for 2 Pro device.
* Reduced lowest brightness level for Pro 2
device.
* Changed the format in which brightness values
are stored in database - previously it was
stored as PWM percentage, now 1-10 scale
system level is used.

+++ products/BellHybrid/apps/application-bell-powernap/models/PowerNapFrontlightModel.hpp
@@ -30,7 +30,7 @@ namespace app::powernap
         const std::chrono::seconds alarmDuration;
         mutable settings::Settings settings;
 
-        auto getAlarmBrightness() const -> int;
+        auto getAlarmBrightness() const -> float;
         auto prepareFadeInParameters() const -> screen_light_control::LinearProgressModeParameters;
     };
 } // namespace app::powernap
